recordes quase completos?

+++ src/main/java/pt/ipleiria/estg/dei/ei/esoft/TabelaRecordes.java
@@ -2,15 +2,14 @@ package pt.ipleiria.estg.dei.ei.esoft;
 
 public class TabelaRecordes {
     private String nome;
-    private int tempo;
-
+    private long tempo;
 
     public TabelaRecordes() {
         this.nome = "Anónimo";
         this.tempo = 9999999;
     }
 
-    public int getTempo() {
+    public long getTempo() {
         return tempo;
     }
 
@@ -32,4 +31,5 @@ public class TabelaRecordes {
             this.nome = nome;
         }
     }
+
 }
